requirements.txt
astroid==1.6.2
bitarray==0.8.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
SimpleWebsocketServerFork
mccabe==0.6.1
plumbum==1.6.6
pylint==1.8.3
PyYAML==3.12
requests==2.18.4
rpyc==3.4.4
Savoir==1.0.6
six==1.11.0
urllib3==1.22
websocket-client==0.46.0
wrapt==1.10.11
